+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added welcome screen.
 - Use current project filter to prefill project field in task form.
 - Change app icon on iOS.
+- Render task text as markdown.
 
 ## 0.3.0